component: dynamic bottom navbar modified
