+++ src/pages/product/detail.jsx
@@ -1,6 +1,6 @@
 
 import React, { useState } from "react"
-import { Badge, Box, Button, Grid, GridItem, Icon, IconButton, Tab, Table, TableContainer, TabList, TabPanel, TabPanels, Tabs, Tbody, Td, Text, Tr, useBreakpointValue } from "@chakra-ui/react"
+import { Badge, Box, Button, Grid, GridItem, Icon, IconButton, Table, TableContainer, Tbody, Td, Text, Tr, useBreakpointValue } from "@chakra-ui/react"
 import { HiPlus, HiMinus, HiShoppingCart, HiOutlineHeart, HiOutlineShare, HiOutlineChat } from "react-icons/hi"
 import Slider from "react-slick"
 import { BiLeftArrowAlt, BiRightArrowAlt } from "react-icons/bi";
@@ -124,193 +124,168 @@ const detail = () => {
 
     return (
         <Grid
+            templateRows={["repeat(3)", "repeat(3, 1fr)", "repeat(3, 1fr)"]}
             templateColumns={["repeat(1, 1fr)", "repeat(5, 1fr)", "repeat(5, 1fr)"]}
-            gap={[2, 2, 2]}
+            gap={[2, 2, 4]}
             alignContent="center"
             w={["100%", "90", "90%"]}
             marginLeft="auto"
             marginRight="auto"
-        // mb={[10, 8, 8]}
+            mb={[10, 8, 8]}
         >
             <GridItem
+                rowSpan={[1, 2, 2]}
                 colSpan={[1, 2, 2]}
                 padding={[2, 6, 6]}
                 alignItems="center"
                 justifyContent="center"
                 display="flex-Box"
-
+                boxShadow="xl"
             >
                 <Box
-                    boxShadow={["none", "base", "base"]}
-                    width="100%"
-                    padding={[2, 6, 6]}
-                    borderRadius="16px"
+                    position="relative"
+                    height={[250, 300, 300]}
+                    width={[300, 405, 405]}
+                    maxW="100%"
+                    overflow="hidden"
+                    alignItems="center"
+                    justifyContent="center"
+                    marginRight="auto"
+                    marginLeft="auto"
                 >
-
-                    <Box
-                        position="relative"
-                        height={[250, 300, 300]}
-                        width={[300, 405, 405]}
-                        maxW="100%"
-                        overflow="hidden"
-                        alignItems="center"
-                        justifyContent="center"
-                        marginRight="auto"
-                        marginLeft="auto"
-
+                    {/* CSS files for react-slick */}
+                    <link
+                        rel="stylesheet"
+                        type="text/css"
+                        charSet="UTF-8"
+                        href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css"
+                    />
+                    <link
+                        rel="stylesheet"
+                        type="text/css"
+                        href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"
+                    />
+                    {/* Left Icon */}
+                    <IconButton
+                        aria-label="left-arrow"
+                        colorScheme="messenger"
+                        borderRadius="full"
+                        position="absolute"
+                        left={side}
+                        top={top}
+                        transform="translate(0%, -50%)"
+                        zIndex={2}
+                        onClick={() => slider?.slickPrev()}
+                        display={["none", "flex"]}
                     >
-                        {/* CSS files for react-slick */}
-                        <link
-                            rel="stylesheet"
-                            type="text/css"
-                            charSet="UTF-8"
-                            href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css"
-                        />
-                        <link
-                            rel="stylesheet"
-                            type="text/css"
-                            href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"
-                        />
-                        {/* Left Icon */}
-                        <IconButton
-                            aria-label="left-arrow"
-                            colorScheme="messenger"
-                            borderRadius="full"
-                            position="absolute"
-                            left={side}
-                            top={top}
-                            transform="translate(0%, -50%)"
-                            zIndex={2}
-                            onClick={() => slider?.slickPrev()}
-                            display={["none", "flex"]}
-                        >
-                            <BiLeftArrowAlt />
-                        </IconButton>
-                        {/* Right Icon */}
-                        <IconButton
-                            aria-label="right-arrow"
-                            colorScheme="messenger"
-                            borderRadius="full"
-                            position="absolute"
-                            right={side}
-                            top={top}
-                            transform="translate(0%, -60%)"
-                            zIndex={2}
-                            onClick={() => slider?.slickNext()}
-                            display={["none", "flex"]}
-                        >
-                            <BiRightArrowAlt />
-                        </IconButton>
-                        {/* Slider */}
-                        <Slider {...settings} ref={() => setSlider(slider)}>
-                            {cards.map((url) => (
-                                <Box
-                                    key={url}
-                                    height={[180, 250, 250]}
-                                    width={["full", 405, 405]}
-                                    position="relative"
-                                    backgroundPosition="center"
-                                    backgroundRepeat="no-repeat"
-                                    backgroundSize="cover"
-                                    backgroundImage={`url(${url})`}
-                                />
-                            ))}
-                        </Slider>
-                    </Box>
-                    <Box mt={2} display={["none", "flex", "flex"]}>
-                        <Button width="145px" height="46px" variant="main" mr={5} borderRadius="130px" alignItems="center"  >
-                            <Icon mr={2} as={HiOutlineChat} />
-                            Chat Admin
-                        </Button>
-                        <Button width="145px" height="46px" variant="main" borderRadius="130px" alignItems="center" >
-                            <Icon mr={2} as={HiOutlineShare} />
-                            Bagikan
-                        </Button>
-                    </Box>
+                        <BiLeftArrowAlt />
+                    </IconButton>
+                    {/* Right Icon */}
+                    <IconButton
+                        aria-label="right-arrow"
+                        colorScheme="messenger"
+                        borderRadius="full"
+                        position="absolute"
+                        right={side}
+                        top={top}
+                        transform="translate(0%, -60%)"
+                        zIndex={2}
+                        onClick={() => slider?.slickNext()}
+                        display={["none", "flex"]}
+                    >
+                        <BiRightArrowAlt />
+                    </IconButton>
+                    {/* Slider */}
+                    <Slider {...settings} ref={() => setSlider(slider)}>
+                        {cards.map((url) => (
+                            <Box
+                                key={url}
+                                height={[180, 250, 250]}
+                                width={["full", 405, 405]}
+                                position="relative"
+                                backgroundPosition="center"
+                                backgroundRepeat="no-repeat"
+                                backgroundSize="cover"
+                                backgroundImage={`url(${url})`}
+                            />
+                        ))}
+                    </Slider>
+                </Box>
 
+                <Box mt={2} display={["none", "flex", "flex"]}>
+                    <Button width="40%" variant="main" mr={5} borderRadius="130px" alignItems="center"  >
+                        <Icon mr={2} as={HiOutlineChat} />
+                        Chat Admin
+                    </Button>
+                    <Button width="40%" variant="main" borderRadius="130px" alignItems="center" >
+                        <Icon mr={2} as={HiOutlineShare} />
+                        Bagikan
+                    </Button>
                 </Box>
             </GridItem>
-            <GridItem hight={["auto", "auto", "auto"]} padding={[2, 6, 6]} colSpan={[1, 3, 3]} >
-                <Box
-                    // boxShadow={["none", "dark-lg", "dark-lg"]}
-                    padding={[2, 6, 6]}
-                    borderRadius="16px"
-                    margin={[2, 4, 4]}
-                >
+            <GridItem hight={["auto", "auto", "auto"]} padding={3} colSpan={[1, 3, 3]} rowSpan={[1, 1, 1]} boxShadow="xl" >
+                <Box>
+                    <Text variant='mini-title'>Bisolvon</Text>
+                    <Text variant="caption-ligth">Bisolvon 8MG 4 Tablet</Text>
+                    <Box display="flex" alignItems='center'>
+                        <Text variant="title">Rp. 13.000</Text> &nbsp;
+                        <Text variant="caption"> / per strip(4 tablet)</Text>
+                    </Box>
+                    <Text as='s' variant="caption">Rp.17.000</Text>
+                    <Badge ml={2}>17%</Badge>
+                </Box>
+                <Box display="flex" alignItems="center">
+                    <IconButton>
+                        <Icon as={HiMinus} />
+                    </IconButton>
 
-                    <Box>
-                        <Text variant='mini-title'>Bisolvon</Text>
-                        <Text variant="caption-ligth">Bisolvon 8MG 4 Tablet</Text>
-                        <Box display="flex" alignItems='center'>
-                            <Text variant="title">Rp. 13.000</Text> &nbsp;
-                            <Text variant="caption"> / per strip(4 tablet)</Text>
-                        </Box>
-                        <Box mt={3} display="flex" alignItems='center'>
-                            <Text as='s' variant="caption">Rp.17.000</Text>
-                            <Badge ml={2}>17%</Badge>
-                        </Box>
-                    </Box>
-                    <Box alignItems="center" display="flex" mt={5}>
-                        <Box display="flex" alignItems="center" justifyContent="space-between" width={["138px", "164px"]} height={["36px", "38px"]}>
-                            <IconButton>
-                                <Icon as={HiMinus} />
-                            </IconButton>
-                            <Box variant="main" textAlign="center">1</Box>
-                            <IconButton>
-                                <Icon as={HiPlus} />
-                            </IconButton>
-                        </Box>
-                        <Text variant="caption-bold" ml={5}> Sisa 20 Strips</Text>
-                    </Box>
-                    {/* <Box marginTop={5}> */}
-                    <Box display={["none", "flex", "flex"]} alignItems="center" mt={5} >
-                        <Button width="194px" height="48px" variant="main">
+                    <Box width="40%" variant="main" margin="5" w="10" textAlign="center">1</Box>
+                    <IconButton>
+                        <Icon as={HiPlus} />
+                    </IconButton>
+                    <Text variant="caption-bold" ml={5}> Sisa 20 Strips</Text>
+                </Box>
+                <Box>
+                    <Box display="flex" alignItems="center"  >
+                        <Button width="30%" variant="main">
                             <Icon mr={2} as={HiShoppingCart} />
                             keranjang
                         </Button >
-                        <Button width="153px" height="48px" variant="main" ml={5}>
+                        <Button width="30%" variant="main" ml={5}>
                             Beli
                         </Button>
                         <IconButton variant="outline" ml={5}>
-                            <Icon size="49px" as={HiOutlineHeart} />
+                            <Icon as={HiOutlineHeart} />
                         </IconButton>
                     </Box>
-                    {/* </Box> */}
-                </Box>
-                <Box
-                    // boxShadow={["none", "dark-lg", "dark-lg"]}
-                    padding={[2, 6, 6]}
-                    borderRadius="16px"
-                    margin={[2, 4, 4]}
-                >
-                    <Tabs
-                        align='center'
-                        // variant="enclosed"
-                        isFitted
-                        isLazy
-                    >
-                        <TabList>
-                            <Tab _focus={{ borderBottomColor: "teal", outline: 0 }}>Deskripsi</Tab>
-                            <Tab _focus={{ borderBottomColor: "teal", outline: 0 }}>Cara Pakai</Tab>
-                            <Tab _focus={{ borderBottomColor: "teal", outline: 0 }}>Peringatan</Tab>
-                        </TabList>
-
-                        <TabPanels>
-                            <TabPanel>
-                                <TableContainer>
-                                    <Table variant='simple'>
-                                        {tabelShow}
-                                    </Table>
-                                </TableContainer>
-                            </TabPanel>
-                            <TabPanel>Cara Pakai</TabPanel>
-                            <TabPanel>Peringatan</TabPanel>
-                        </TabPanels>
-                    </Tabs>
                 </Box>
             </GridItem>
-
-
+            <GridItem padding={3} colSpan={[1, 3, 3]} rowSpan={[1, 2, 2]} boxShadow="xl" >
+                <Box
+                    display="flex"
+                    alignItems="center"
+                    justifyContent="space-evenly"
+                    borderTop="1px solid black"
+                    borderBottom="1px solid black"
+                    height="50px"
+                >
+                    <Box as="button">
+                        Deskripsi
+                    </Box>
+                    <Box as="button">
+                        Cara Pakai
+                    </Box>
+                    <Box as="button">
+                        Peringatan
+                    </Box>
+                </Box>
+                <TableContainer>
+                    <Table variant='simple'>
+                        {tabelShow}
+                    </Table>
+                </TableContainer>
+            </GridItem>
+            <Box />
         </Grid >
     )
 }
